Improve parsing some option in dhcpcd.conf
All DHCP options have a limit of 255 bytes, so just define that
rather than each option we want to parse.
Only the hostname option is NUL terminated, the others store an
extra leading byte to state the data length.
Process the to the size of the storage, not a predifined length
incase we change it again in the future.

More for #435.
Note that for parse_hwaddr, we set the first byte to zero if
we cannot process arg as a hardware address to indicate that
the data should be treated as a string.