route lifetime behaviour should mirror other lifetimes
Each lifetime has a an aquistion time so we can compare to
now to know how much of the lifetime is left.
Apply the difference when adding/changing the route just as we
do for addresses.

This allow us to work out better when we need to update the
route lifetime from a newer aquisition.

Fixes #465.